+++ __tests__/ApplicationTest.js
@@ -1,7 +1,4 @@
 import App from "../src/App.js";
-import Car from "../src/models/Car.js";
-import RacingGame from "../src/models/RacingGame.js";
-import Validator from "../src/utils/Validator.js";
 import { MissionUtils } from "@woowacourse/mission-utils";
 
 const mockQuestions = (inputs) => {
@@ -59,126 +56,5 @@ describe("자동차 경주", () => {
 
     // then
     await expect(app.run()).rejects.toThrow("[ERROR]");
-  });
-});
-
-describe("Validator 클래스 테스트", () => {
-  test("이름이 5자를 초과하면 에러 발생", () => {
-    expect(() => Validator.validateCarName("javajigi")).toThrow("[ERROR]");
-  });
-
-  test("이름이 공백이면 에러 발생", () => {
-    expect(() => Validator.validateCarName("")).toThrow("[ERROR]");
-  });
-
-  test("이름이 공백 문자로만 이루어진 경우 에러 발생", () => {
-    expect(() => Validator.validateCarName("  ")).toThrow("[ERROR]");
-  });
-
-  test("중복된 이름이 있으면 에러 발생", () => {
-    expect(() => Validator.validateDuplicateCarName(["pobi", "pobi"])).toThrow(
-      "[ERROR]"
-    );
-  });
-
-  test("유효한 시도 횟수 입력", () => {
-    expect(Validator.validateTryCount("5")).toBe(5);
-  });
-});
-
-describe("Car 클래스 테스트", () => {
-  test("유효한 이름으로 자동차 생성", () => {
-    const car = new Car("pobi");
-    expect(car.name).toBe("pobi");
-  });
-
-  test("4 이상의 숫자에서 자동차 전진", () => {
-    const car = new Car("pobi");
-    car.moveCar(4);
-    expect(car.position).toBe(1);
-  });
-
-  test("3 이하의 숫자에서 자동차 정지", () => {
-    const car = new Car("pobi");
-    car.moveCar(3);
-    expect(car.position).toBe(0);
-  });
-
-  test("현재 상태 출력 테스트", () => {
-    const car = new Car("pobi");
-    car.moveCar(4);
-    car.moveCar(7);
-    expect(car.getCurrentStatus()).toBe("pobi : --");
-  });
-});
-
-describe("RacingGame 클래스 테스트", () => {
-  test("쉼표로 구분된 이름으로 자동차들 생성", () => {
-    const game = new RacingGame();
-    expect(() => {
-      game.createCars("pobi, jini");
-    }).not.toThrow("[ERROR]");
-  });
-
-  test("중복된 이름의 자동차들 생성 시 에러 발생", () => {
-    const game = new RacingGame();
-    expect(game.validateTryCount("5")).toBe(5);
-  });
-
-  test("유효한 시도 횟수 입력", () => {
-    const game = new RacingGame();
-    expect(() => {
-      game.validateTryCount("-1");
-    }).toThrow("[ERROR]");
-  });
-
-  test("시도 횟수가 0이하일 경우 에러 발생", () => {
-    const game = new RacingGame();
-    expect(() => {
-      game.validateTryCount("-1");
-    }).toThrow("[ERROR]");
-  });
-
-  test("시도 횟수가 숫자가 아닐 경우 에러 발생", () => {
-    const game = new RacingGame();
-    expect(() => {
-      game.validateTryCount("lolol");
-    }).toThrow("[ERROR]");
-  });
-
-  test("모든 자동차 이동 테스트", () => {
-    const game = new RacingGame();
-    game.createCars("pobi, jini");
-
-    mockRandoms([6, 3]);
-
-    game.moveAllCars();
-
-    expect(game.cars[0].position).toBe(1);
-    expect(game.cars[1].position).toBe(0);
-  });
-
-  test("단독 우승자 테스트", () => {
-    const game = new RacingGame();
-    game.createCars("pobi, jini, cho");
-
-    mockRandoms([6, 3, 1, 6, 3, 1]);
-
-    game.moveAllCars();
-
-    const winner = game.findWinners();
-    expect(winner).toEqual(["pobi"]);
-  });
-
-  test("공동 우승자 테스트", () => {
-    const game = new RacingGame();
-    game.createCars("pobi, jini, cho");
-
-    mockRandoms([4, 4, 3, 4, 4, 2]);
-
-    game.moveAllCars();
-
-    const winners = game.findWinners();
-    expect(winners).toEqual(["pobi", "jini"]);
   });
 });
